# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": { "title": "Full Stack FastAPI Project", "version": "0.1.0" },
  "paths": {
...
    "/api/v1/users/": {
      "get": {
        "tags": ["users"],
        "summary": "Read Users",
        "description": "Retrieve users.",
        "operationId": "users-read_users",
        "security": [{ "SupabaseBearer": [] }],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "default": 0, "title": "Skip" }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "default": 100, "title": "Limit" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/UsersPublic" } }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["users"],
        "summary": "Create User",
        "description": "Create new user.",
        "operationId": "users-create_user",
        "security": [{ "SupabaseBearer": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/UserCreate" } }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/UserPublic" } }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/me": {
      "get": {
        "tags": ["users"],
        "summary": "Read User Me",
        "description": "Get current user.",
        "operationId": "users-read_user_me",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/UserPublic" } }
            }
          }
        },
        "security": [{ "SupabaseBearer": [] }]
      },
      "delete": {
        "tags": ["users"],
        "summary": "Delete User Me",
        "description": "Delete own user.",
        "operationId": "users-delete_user_me",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Message" } }
            }
          }
        },
        "security": [{ "SupabaseBearer": [] }]
      },
      "patch": {
        "tags": ["users"],
        "summary": "Update User Me",
        "description": "Update own user.",
        "operationId": "users-update_user_me",
        "requestBody": {
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/UserUpdateMe" } }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/UserPublic" } }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        },
        "security": [{ "SupabaseBearer": [] }]
      }
    },
    "/api/v1/users/me/avatar": {
      "post": {
        "tags": ["users"],
        "summary": "Upload Avatar Me",
        "description": "Upload avatar for the current user.\n\nThe avatar will be uploaded to Supabase Storage in the 'avatars' bucket\nand the user's avatar_url will be updated in the database.\n\nExpected: multipart/form-data with 'avatar' field containing the image file.\nMax size: 5MB. Allowed types: image/jpeg, image/png, image/webp.",
        "operationId": "users-upload_avatar_me",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": { "$ref": "#/components/schemas/Body_users-upload_avatar_me" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/UserPublic" } }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        },
        "security": [{ "SupabaseBearer": [] }]
      }
    },
    "/api/v1/users/{user_id}": {
      "get": {
        "tags": ["users"],
        "summary": "Read User By Id",
        "description": "Get a specific user by id.",
        "operationId": "users-read_user_by_id",
        "security": [{ "SupabaseBearer": [] }],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid", "title": "User Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/UserPublic" } }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["users"],
        "summary": "Update User",
        "description": "Update a user.",
        "operationId": "users-update_user",
        "security": [{ "SupabaseBearer": [] }],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid", "title": "User Id" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/UserUpdate" } }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/UserPublic" } }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["users"],
        "summary": "Delete User",
        "description": "Delete a user.",
        "operationId": "users-delete_user",
        "security": [{ "SupabaseBearer": [] }],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid", "title": "User Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Message" } }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/services/": {
      "get": {
        "tags": ["services"],
        "summary": "Read Services",
        "description": "Retrieve services.",
        "operationId": "services-read_services",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ServicesPublic" } }
            }
          }
        }
      },
      "post": {
        "tags": ["services"],
        "summary": "Create Service Endpoint",
        "description": "Create new service.",
        "operationId": "services-create_service_endpoint",
        "requestBody": {
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/ServiceCreate" } }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ServicePublic" } }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/services/{id}": {
      "get": {
        "tags": ["services"],
        "summary": "Read Service",
        "description": "Get service by ID.",
        "operationId": "services-read_service",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid", "title": "Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ServicePublic" } }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["services"],
        "summary": "Update Service",
        "description": "Update a service.",
        "operationId": "services-update_service",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid", "title": "Id" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/ServiceUpdate" } }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ServicePublic" } }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["services"],
        "summary": "Delete Service Endpoint",
        "description": "Delete a service.",
        "operationId": "services-delete_service_endpoint",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid", "title": "Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": { "type": "string" },
                  "title": "Response Services-Delete Service Endpoint"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/private/users/": {
      "post": {
        "tags": ["private"],
        "summary": "Create User",
        "description": "Create a new user.",
        "operationId": "private-create_user",
        "requestBody": {
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/PrivateUserCreate" } }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/UserPublic" } }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Body_users-upload_avatar_me": {
        "properties": { "avatar": { "type": "string", "format": "binary", "title": "Avatar" } },
        "type": "object",
        "required": ["avatar"],
        "title": "Body_users-upload_avatar_me"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": { "$ref": "#/components/schemas/ValidationError" },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Message": {
        "properties": { "message": { "type": "string", "title": "Message" } },
        "type": "object",
        "required": ["message"],
        "title": "Message"
      },
...
      "PrivateUserCreate": {
        "properties": {
          "email": { "type": "string", "title": "Email" },
          "full_name": { "type": "string", "title": "Full Name" },
          "is_verified": { "type": "boolean", "title": "Is Verified", "default": false }
        },
        "type": "object",
        "required": ["email", "full_name"],
        "title": "PrivateUserCreate"
      },
...
      "ServiceCreate": {
        "properties": {
          "category": { "type": "string", "maxLength": 255, "minLength": 1, "title": "Category" },
          "description": {
            "anyOf": [{ "type": "string", "maxLength": 255 }, { "type": "null" }],
            "title": "Description"
          },
          "duration": { "type": "integer", "minimum": 0.0, "title": "Duration" },
          "price": { "type": "number", "minimum": 0.0, "title": "Price" }
        },
        "type": "object",
        "required": ["category", "duration", "price"],
        "title": "ServiceCreate"
      },
      "ServicePublic": {
        "properties": {
          "category": { "type": "string", "maxLength": 255, "minLength": 1, "title": "Category" },
          "description": {
            "anyOf": [{ "type": "string", "maxLength": 255 }, { "type": "null" }],
            "title": "Description"
          },
          "duration": { "type": "integer", "minimum": 0.0, "title": "Duration" },
          "price": { "type": "number", "minimum": 0.0, "title": "Price" },
          "id": { "type": "string", "format": "uuid", "title": "Id" }
        },
        "type": "object",
        "required": ["category", "duration", "price", "id"],
        "title": "ServicePublic"
      },
      "ServiceUpdate": {
        "properties": {
          "category": { "type": "string", "maxLength": 255, "minLength": 1, "title": "Category" },
          "description": {
            "anyOf": [{ "type": "string", "maxLength": 255 }, { "type": "null" }],
            "title": "Description"
          },
          "duration": { "type": "integer", "minimum": 0.0, "title": "Duration" },
          "price": { "type": "number", "minimum": 0.0, "title": "Price" }
        },
        "type": "object",
        "title": "ServiceUpdate"
      },
      "ServicesPublic": {
        "properties": {
          "data": {
            "items": { "$ref": "#/components/schemas/ServicePublic" },
            "type": "array",
            "title": "Data"
          },
          "count": { "type": "integer", "title": "Count" }
        },
        "type": "object",
        "required": ["data", "count"],
        "title": "ServicesPublic"
      },
      "UserCreate": {
        "properties": {
          "email": { "type": "string", "maxLength": 255, "format": "email", "title": "Email" },
          "is_active": { "type": "boolean", "title": "Is Active", "default": true },
          "is_superuser": { "type": "boolean", "title": "Is Superuser", "default": false },
          "full_name": {
            "anyOf": [{ "type": "string", "maxLength": 255 }, { "type": "null" }],
            "title": "Full Name"
          },
          "name": { "anyOf": [{ "type": "string" }, { "type": "null" }], "title": "Name" },
          "about": { "anyOf": [{ "type": "string" }, { "type": "null" }], "title": "About" },
          "avatar_url": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Avatar Url"
          }
        },
        "type": "object",
        "required": ["email"],
        "title": "UserCreate"
      },
      "UserPublic": {
        "properties": {
          "email": { "type": "string", "maxLength": 255, "format": "email", "title": "Email" },
          "is_active": { "type": "boolean", "title": "Is Active", "default": true },
          "is_superuser": { "type": "boolean", "title": "Is Superuser", "default": false },
          "full_name": {
            "anyOf": [{ "type": "string", "maxLength": 255 }, { "type": "null" }],
            "title": "Full Name"
          },
          "name": { "anyOf": [{ "type": "string" }, { "type": "null" }], "title": "Name" },
          "about": { "anyOf": [{ "type": "string" }, { "type": "null" }], "title": "About" },
          "avatar_url": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Avatar Url"
          },
          "id": { "type": "string", "format": "uuid", "title": "Id" }
        },
        "type": "object",
        "required": ["email", "id"],
        "title": "UserPublic"
      },
      "UserUpdate": {
        "properties": {
          "email": {
            "anyOf": [
              { "type": "string", "maxLength": 255, "format": "email" },
              { "type": "null" }
            ],
            "title": "Email"
          },
          "is_active": { "type": "boolean", "title": "Is Active", "default": true },
          "is_superuser": { "type": "boolean", "title": "Is Superuser", "default": false },
          "full_name": {
            "anyOf": [{ "type": "string", "maxLength": 255 }, { "type": "null" }],
            "title": "Full Name"
          },
          "name": { "anyOf": [{ "type": "string" }, { "type": "null" }], "title": "Name" },
          "about": { "anyOf": [{ "type": "string" }, { "type": "null" }], "title": "About" },
          "avatar_url": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Avatar Url"
          },
          "password": {
            "anyOf": [{ "type": "string", "maxLength": 40, "minLength": 8 }, { "type": "null" }],
            "title": "Password"
          }
        },
        "type": "object",
        "title": "UserUpdate"
      },
      "UserUpdateMe": {
        "properties": {
          "full_name": {
            "anyOf": [{ "type": "string", "maxLength": 255 }, { "type": "null" }],
            "title": "Full Name"
          },
          "email": {
            "anyOf": [
              { "type": "string", "maxLength": 255, "format": "email" },
              { "type": "null" }
            ],
            "title": "Email"
          },
          "name": { "anyOf": [{ "type": "string" }, { "type": "null" }], "title": "Name" },
          "about": { "anyOf": [{ "type": "string" }, { "type": "null" }], "title": "About" },
          "avatar_url": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Avatar Url"
          }
        },
        "type": "object",
        "title": "UserUpdateMe"
      },
      "UsersPublic": {
        "properties": {
          "data": {
            "items": { "$ref": "#/components/schemas/UserPublic" },
            "type": "array",
            "title": "Data"
          },
          "count": { "type": "integer", "title": "Count" }
        },
        "type": "object",
        "required": ["data", "count"],
        "title": "UsersPublic"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
            "type": "array",
            "title": "Location"
          },
          "msg": { "type": "string", "title": "Message" },
          "type": { "type": "string", "title": "Error Type" }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      }
    },
    "securitySchemes": { "SupabaseBearer": { "type": "http", "scheme": "bearer" } }
  }
}